+++ llvm-hs/test/LLVM/Test/Linking.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 module LLVM.Test.Linking where
 
 import Test.Tasty