+++ llvm-hs/test/LLVM/Test/Metadata.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 module LLVM.Test.Metadata where
 
 import Test.Tasty